Reorganize server folder structure
