--- java-server/src/main/java/com/example/game/game/GameConstants.java
@@ -1,10 +1,0 @@
-package com.example.game.game;
-
-public final class GameConstants {
-    public static final int WIDTH = 800;
-    public static final int HEIGHT = 500;
-
-    public static final int TICK_DELAY_MS = 20;
-
-    private GameConstants() {}
-}
