--- java-server/src/main/java/com/example/game/players/PlayerControl.java
@@ -1,17 +1,0 @@
-package com.example.game.players;
-
-import com.fasterxml.jackson.annotation.JsonValue;
-
-public enum PlayerControl {
-    UP("Up"),
-    DOWN("Down"),
-    LEFT("Left"),
-    RIGHT("Right");
-
-    @JsonValue
-    public final String jsonValue;
-
-    PlayerControl(String value) {
-        this.jsonValue = value;
-    }
-}
