--- java-server/src/main/java/com/example/game/players/PlayerState.java
@@ -1,15 +1,0 @@
-package com.example.game.players;
-
-public record PlayerState(
-    /** CSS compatible string */
-    String color,
-
-    /** X position */
-    int x,
-
-    /** Y Position */
-    int y,
-
-    /** Number of seconds since the player has joined */
-    int age
-) {}
