--- java-server/src/main/java/com/example/game/sessions/Singletons.java
@@ -1,9 +1,0 @@
-package com.example.game.sessions;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-public final class Singletons {
-    public static final ObjectMapper objectMapper = new ObjectMapper();
-
-    private Singletons() {}
-}
