--- java-server/src/test/java/com/example/game/AppTest.java
@@ -1,16 +1,0 @@
-package com.example.game;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AppTest {
-
-	@Test
-	void test_junitTests_run() {
-		assertEquals(1 + 1, 2);
-	}
-
-}
