--- java-server/src/test/java/com/example/game/GameApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.game;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class GameApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
