+++ server/src/main/java/io/github/aggarcia/game/GameConstants.java
@@ -1,0 +1,10 @@
+package io.github.aggarcia.game;
+
+public final class GameConstants {
+    public static final int WIDTH = 800;
+    public static final int HEIGHT = 500;
+
+    public static final int TICK_DELAY_MS = 20;
+
+    private GameConstants() {}
+}
