+++ server/src/main/java/io/github/aggarcia/players/PlayerControl.java
@@ -1,0 +1,17 @@
+package io.github.aggarcia.players;
+
+import com.fasterxml.jackson.annotation.JsonValue;
+
+public enum PlayerControl {
+    UP("Up"),
+    DOWN("Down"),
+    LEFT("Left"),
+    RIGHT("Right");
+
+    @JsonValue
+    public final String jsonValue;
+
+    PlayerControl(String value) {
+        this.jsonValue = value;
+    }
+}
