+++ server/src/main/java/io/github/aggarcia/players/PlayerControlUpdate.java
@@ -1,0 +1,10 @@
+package io.github.aggarcia.players;
+
+import java.util.List;
+
+import io.github.aggarcia.shared.SocketMessage;
+
+public record PlayerControlUpdate(
+    SocketMessage type,
+    List<PlayerControl> pressedControls
+) {}
