+++ server/src/main/java/io/github/aggarcia/shared/SocketMessage.java
@@ -1,0 +1,17 @@
+package io.github.aggarcia.shared;
+
+import com.fasterxml.jackson.annotation.JsonValue;
+
+public enum SocketMessage {
+    GAME_UPDATE("gameUpdate"),
+    GAME_OVER_UPDATE("gameOverUpdate"),
+    PLAYER_CONTROL_UPDATE("playerControlUpdate"),
+    SERVER_ERROR("serverError");
+
+    @JsonValue
+    public final String jsonValue;
+
+    SocketMessage(String value) {
+        this.jsonValue = value;
+    }
+}
